PersonaGen/requirements.txt
pydantic>=2.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.1
PyYAML>=6.0
python-dotenv>=1.0.0
